✨ feat: Add fileManager class with methods for downloading and unzipping files
This commit adds a new class called fileManager with methods for downloading, unzipping, and getting the download link of a file. It also includes changes to package.json to add a dependency on unzip-stream and @types/cheerio. Finally, it adds two new files: index.ts (which currently just exports an empty object) and parser.ts (which imports the fileManager class). Additionally, it updates the .gitignore file to include file.zip.